Remove item that represents a bad practice
Pushing to production on a Friday is potentially risky for the health of the project, can inconvenience customers, and can be considered disrespectful to the other engineers who may be called to fix things over the weekend. It doesn’t seem to belong in this list of otherwise positive, constructive qualities.