Dear VC,
When we felt our values threatened by your tweet, we did a good thing and a bad
thing. The good thing: we responded collaboratively by creating this project,
which strives to honor the humanity of engineers all around the world. We are
proud to stand up for our values in this way.

The bad thing we did is that we reacted by name-calling. We called you clueless,
and that was not kind. We are sorry. Will you please forgive us?

Warm regards,
1x Engineer