Indiscriminate Copy / Paste is not a good.
From this fragment, it looks like copy & pastes with no control is good behavior, when is the main problem of a lot of engineers and of course it is not part of the 1x engineer's daily basics.
Copy & Paste must be along with understanding.

Also, it is the second point. So it looks like copy & paste is one of the more important things when in fact, it is one of the more dangerous things.