feat: Stands up for others
We've all found ourselves in those calls/meetings where a single individual has wrongly/unfairly been assigned all the blame for some particular failure and the criticism is exceeding the levels acceptable in a professional environment. The 1x engineer stands up to defend the accused at potential risk to themselves.